+++ libtwolame/psycho_2.c
@@ -251,8 +251,8 @@ psycho_2_mem *psycho_2_init(twolame_options * glopts, int sfreq)
     return (mem);
 }
 
-void psycho_2(twolame_options * glopts, short int buffer[2][1152],
-              short int savebuf[2][1056], FLOAT smr[2][32])
+void psycho_2(twolame_options * glopts, int32_t buffer[2][1152],
+              int32_t savebuf[2][1056], FLOAT smr[2][32])
 {
     psycho_2_mem *mem;
     unsigned int i, j, k, ch;
@@ -328,7 +328,7 @@ void psycho_2(twolame_options * glopts, short int buffer[2][1152],
 		   BLKSIZE = 1024
 	   *****************************************************************************/
             {
-                short int *bufferp = buffer[ch];
+                int32_t *bufferp = buffer[ch];
                 for (j = 0; j < 480; j++) {
                     savebuf[ch][j] = savebuf[ch][j + mem->flush];
                     wsamp_r[j] = window[j] * ((FLOAT) savebuf[ch][j]);
